requirements.txt
unsloth
bitsandbytes
accelerate
xformers==0.0.29.post3
peft
trl==0.15.2
torch
triton
cut_cross_entropy
unsloth_zoo
sentencepiece
protobuf
datasets
huggingface_hub
hf_transfer
Pillow
python-dotenv
wandb
pyyaml
tqdm